refactor: move gateway api structs
