+++ gateway/ln-gateway/src/lib.rs
@@ -67,9 +67,9 @@ use fedimint_core::{fedimint_build_code_version_env, Amount, BitcoinAmountOrAll}
 use fedimint_ln_common::config::{GatewayFee, LightningClientConfig};
 use fedimint_ln_common::contracts::Preimage;
 use fedimint_ln_common::LightningCommonInit;
+use fedimint_lnv2_client::api::{CreateBolt11InvoicePayload, SendPaymentPayload};
 use fedimint_lnv2_client::{
-    Bolt11InvoiceDescription, CreateBolt11InvoicePayload, PaymentFee, RoutingInfo,
-    SendPaymentPayload, EXPIRATION_DELTA_LIMIT_DEFAULT,
+    Bolt11InvoiceDescription, PaymentFee, RoutingInfo, EXPIRATION_DELTA_LIMIT_DEFAULT,
 };
 use fedimint_lnv2_common::contracts::{IncomingContract, PaymentImage};
 use fedimint_mint_client::{
